requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.110.0
uvicorn==0.27.1

torch==2.2.2
torchvision==0.17.2

numpy==1.26.4
pillow==10.2.0

python-multipart==0.0.9